Topic: small patch in validator and more details on previous commit (9064a01)
This commit corrects a small error in the `Execute` endpoint of the validator.
Previously, the third argument of the function unwrapped the `txInfoValidRange` instead the `txInfoRedeemers` used in the endpoint. Already fixed.

The commit https://github.com/WolframBlockchainLabs/MarloweSC-Execution/commit/9064a0189f5adf1187f5c9b9701e9ba76282a639 includes changes not documented in the commit message.
More details in the following lines.
* Changes in library imports to avoid abiguity.
* ScriptContext custom view implemented. This deserializes only the needed elements and left other as BuiltinData.
* Branched validator behaviour. Now validation depends on the redeemer.
* New compilation step, the Tweag audit included a custom wrapper to convert the Typed validator into an Untyped one.
* writeOracleValidator function now writes the new instance of the oracle.